+++ composables/useFoodStarted.ts
@@ -1,6 +1,0 @@
-export const useFoodStarted = () => {
-  const indicators = useFoodIndicators();
-  return computed(() => {
-    return Object.getOwnPropertyNames(indicators.value).length > 1;
-  });
-}